Create README.md
This is the final project for my python course at Bentley University. Throughout the project, I used streamlit, pandas and matplotlib to show the data in a more visualize way. The data set I chose was Boston Crimes, I chose 9 groups of data from the dataset among with dates, types and districts. With the pandas and streamlit, I created the table pie chart and the map.